% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_alpha_diversity.R
\name{boxplot_alpha_diversity}
\alias{boxplot_alpha_diversity}
\title{Create alpha diversity plots}
\usage{
boxplot_alpha_diversity(
  objPhyloseq,
  orden,
  xFactor,
  colFactor,
  divMeasure,
  wrap
)
}
\arguments{
\item{objPhyloseq}{a phyloseq object.}
}
\description{
This fuction takes a phyloseq object and create a pdf file with plots information.
}
\details{
This function is part of a package used for the analysis of microbial communities.
}
\examples{
boxplot_alpha_diversity(mergedata, o, "SampleType", "SampleType", "Simpson", "Host")
}
